contrib/mongo-c-driver/.evergreen/config_generator/requirements.txt
-e git+https://github.com/mongodb-labs/drivers-evergreen-tools#egg=evergreen_config_generator&subdirectory=evergreen_config_generator
git-url-parse==1.2.2
packaging==23.1
pbr==5.11.1
pydantic==1.10.4
PyYAML==5.4.1
shrub.py==3.0.4
typing_extensions==4.4.0
yamlordereddictloader==0.4.0
